refactor: import auth, checkers and uuid directly
These modules were referenced indirectly: cowrie.core.auth via a
sys.modules string lookup, and cowrie.core.checkers/uuid imported by
submodule but accessed through a separate 'core' binding. Pyright could
not see the references and reported them as unused imports.

Import the names directly so the usages are visible to static analysis
and the indirection is removed.
